Introduce CustomResultRepoClient
